github/migrate-custom-props/requirements.txt
argcomplete==3.6.3
astroid==2.5.8
bandit==1.7.5
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==4.10.0
coverage==5.5
cryptography==46.0.3
decli==0.6.3
Deprecated==1.3.1
exceptiongroup==1.3.0
flake8==7.0.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.8
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
lazy-object-proxy==1.10.0
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.7.1
mypy_extensions==1.1.0
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.2.0
PyGithub==2.8.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pylint==2.7.2
PyNaCl==1.6.1
pytest==7.4.3
pytest-cov==4.1.0
python-coveralls==2.9.3
python-dotenv==1.0.0
PyYAML==6.0.2
questionary==2.1.1
requests==2.32.3
rich==14.2.0
smmap==5.0.2
stevedore==5.5.0
termcolor==3.1.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.2.2
wcwidth==0.2.14
wrapt==1.12.1
zipp==3.23.0
